refactor(zk-explorer): Break down into multiple components
